added naming convention checks to clang-tidy
Added tests for function traits

Revert "added naming convention checks to clang-tidy"

This reverts commit d62368db3269a8eaa4c9ea4f4b681e3195102f86.

Added more tests for function traits.
Added generateTrompiline function, that acts the same way as
castMethodToFunction, but can have any input type, that can be
static_cast'ed to the given class. (void* is default)
added overload to castMethodToFunction, so that it is easier to work
with overloaded methods(Signature and Class should be passed as 2 first
template parameters)